single_node_app_with_kafka_compose/pyproject.toml
[tool.poetry]
name = "kafka_compose_single_broker"
version = "0.1.0"
description = ""
authors = ["Muhammad Saad <saadjameel2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
oauth = "^1.0.1"
pycryptodome = "^3.20.0"
python-jose = "^3.3.0"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
